Make adskip optional

+++ spotify.sh
@@ -1,11 +1,16 @@
 #!/usr/bin/env bash
 
+# option to skip advertisements
+# might be visually disruptive if spotify isn't bound to a currently unused workspace
+$ADSKIP=1
+
 # read dbus-monitor 
 dbus-monitor --profile "interface=org.freedesktop.DBus.Properties,member=PropertiesChanged" | while read -r line; do
+
     # parse status
     STATUS=$(dbus-send --print-reply --dest=org.mpris.MediaPlayer2.spotify /org/mpris/MediaPlayer2 org.freedesktop.DBus.Properties.Get string:'org.mpris.MediaPlayer2.Player' string:'PlaybackStatus' | tail -n 1 | sed 's/.*"\(.*\)".*/\1/g') 
 
-    # change toggle icon
+    # change toggle icon based on status
     polybar-msg hook spotify-control $([ "$STATUS" = "Playing" ] && echo 1 || echo 2) | 2&>1 /dev/null
 
     # parse metadata
@@ -14,7 +19,7 @@ dbus-monitor --profile "interface=org.freedesktop.DBus.Properties,member=Propert
     TITLE=$(echo "$METADATA" | grep -A1 "xesam:title" | tail -n 1 | sed 's/.*"\(.*\)".*/\1/g')
 
     # sneaky adblock
-    if [ "$TITLE" = "Advertisement" ] || [ "$TITLE" = "Spotify" ]; then
+    if [ "$ADSKIP" = 1 ] && [[ "$TITLE" = "Advertisement" || "$TITLE" = "Spotify" ]]; then
         pkill spotify
         spotify 2&>1 /dev/null &
         sleep 4
